moving logic from components to hooks
