+++ src/components/text/Timer.jsx
@@ -1,14 +1,8 @@
-import { useEffect, useState } from "react";
+import { useTimer } from "../../hooks";
 
 export function Timer()
 {
-    const [time, setTime] = useState(60);
-    useEffect(() => { 
-        const interval = setInterval(() => { 
-            setTime(prev => (prev - 1 < 0 ? prev : prev - 1)) 
-        }, 1000); 
-        return () => clearInterval(interval); 
-    }, []);
+    const time = useTimer();
     return <>
         <p className="defaultText">timer: {time}s</p>
     </>
