+++ src/hooks/engine/useRandomNumber.jsx
@@ -1,0 +1,4 @@
+export function useRandomNumber({ level })
+{
+    return Math.floor(Math.random() * (level === 'easy' ? 100 : 1000));
+}
